requirements.txt
pandas
streamlit
tensorflow
opencv-python
pillow
numpy